LM/requirements.txt
openai
supabase
Pillow
